+++ monitors/onlinemon/src/TGraphSet.cc
@@ -48,7 +48,7 @@ void TGraphSet::AddPoint(const std::string name, const double time,
     _yMax = value + 0.1;
     _xMin = time - 0.1;
     _xMax = time + 0.1;
-    SetLimits(time - 0.1, time + 0.1, value - 0.1, value + 0.1);
+    SetLimits(static_cast<int>(time - 0.1), static_cast<int>(time + 0.1), static_cast<int>(value - 0.1), static_cast<int>(value + 0.1));
     _starttime = time;
     _nodata = false;
   } else {
